fix: well-known route
